@@ -14,36 +14,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.1.1",
-    "@angular/common": "4.1.1",
-    "@angular/compiler": "4.1.1",
-    "@angular/core": "4.1.1",
-    "@angular/forms": "4.1.1",
-    "@angular/http": "4.1.1",
-    "@angular/platform-browser": "4.1.1",
-    "@angular/platform-browser-dynamic": "4.1.1",
-    "@angular/router": "4.1.1",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.22",
+    "@agm/core": "1.0.0-beta.0",
+    "@angular/animations": "4.3.5",
+    "@angular/cdk": "^2.0.0-beta.10",
+    "@angular/common": "4.3.5",
+    "@angular/compiler": "4.3.5",
+    "@angular/core": "4.3.5",
+    "@angular/flex-layout": "^2.0.0-beta.9",
+    "@angular/forms": "4.3.5",
+    "@angular/http": "4.3.5",
+    "@angular/material": "^2.0.0-beta.10",
+    "@angular/platform-browser": "4.3.5",
+    "@angular/platform-browser-dynamic": "4.3.5",
+    "@angular/router": "4.3.5",
     "@ngx-translate/core": "^6.0.1",
     "@ngx-translate/http-loader": "0.0.3",
-    "angular2-google-maps": "^0.17.0",
     "angular2-jwt": "0.2.3",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
     "crypto-js": "^3.1.9-1",
+    "file-saver": "^1.3.3",
+    "hammerjs": "^2.0.8",
     "intl": "^1.2.5",
+    "js-marker-clusterer": "^1.0.0",
     "ng2-charts": "^1.5.0",
-    "ngx-bootstrap": "1.6.6",
-    "ngx-toastr": "5.2.1",
-    "rxjs": "5.3.1",
+    "ng2-file-upload": "^1.2.1",
+    "ngx-color-picker": "^4.3.1",
+    "node-sass": "^4.5.3",
+    "rxjs": "5.4.2",
     "ts-helpers": "1.1.2",
     "underscore": "^1.8.3",
     "web-animations-js": "2.2.5",
-    "zone.js": "0.8.10"
+    "zone.js": "0.8.17"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.2",
-    "@angular/compiler-cli": "4.1.1",
+    "@angular/cli": "1.3.1",
+    "@angular/compiler-cli": "4.3.5",
+    "@types/file-saver": "0.0.1",
     "@types/jasmine": "2.5.47",
     "@types/node": "7.0.18",
     "codelyzer": "3.0.1",
@@ -55,9 +62,10 @@
     "karma-coverage-istanbul-reporter": "1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "markerwithlabel": "^2.0.0",
     "protractor": "~5.1.0",
-    "ts-node": "3.0.4",
-    "tslint": "5.2.0",
-    "typescript": "2.3.2"
+    "ts-node": "3.3.0",
+    "tslint": "5.7.0",
+    "typescript": "2.5.2"
   }
 }
