@@ -1,8 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { Observable } from 'rxjs/Observable';
+import { Observable, Subject, BehaviorSubject } from 'rxjs/Rx';
 import { Subscription } from 'rxjs/Subscription';
+import { DataSource } from '@angular/cdk/table';
+import { MdPaginator, MdSort } from '@angular/material';
+import { MdDialog } from '@angular/material';
+
 import 'rxjs/add/operator/switchMap';
 import 'rxjs/add/operator/debounceTime';
 import 'rxjs/add/operator/distinctUntilChanged';
@@ -12,14 +15,15 @@ import 'rxjs/add/operator/catch';
 import { LanguagesService } from './languages.service';
 import { LanguageModel } from './language-model';
 import { environment } from '../../../environments/environment';
-import { BaseListComponent, ListFormParams } from '../../core/base-list-component';
+import { BaseListComponent } from '../../core/base-list-component';
+import { ListFormParams } from '../../core/list-form-params';
 
 @Component({
   selector: 'app-languages',
   templateUrl: './languages.component.html',
   styleUrls: ['./languages.component.scss']
 })
-export class LanguagesComponent extends BaseListComponent<LanguageModel> implements OnInit {
+export class LanguagesComponent extends BaseListComponent<LanguageModel, ListFormParams> implements OnInit {
 
   /**
    * Initializes a new instance of the LanguagesComponent.
@@ -31,9 +35,9 @@ export class LanguagesComponent extends BaseListComponent<LanguageModel> impleme
   constructor(
     private languagesService: LanguagesService,
     formBuilder: FormBuilder,
-    modalService: NgbModal
+    public confirmDialog: MdDialog,
   ) {
-    super(formBuilder, modalService);
+    super(formBuilder, confirmDialog);
   }
 
   /**
@@ -41,9 +45,9 @@ export class LanguagesComponent extends BaseListComponent<LanguageModel> impleme
    * @method
    */
   ngOnInit() {
-    this.searchEnabled = false;
-    this.sorts.push({ column: 'name', direction: 'asc' });
     super.ngOnInit();
+    this.dataSource._sort.active = 'name';
+    this.dataSource._sort.direction = 'asc';
   }
 
   /**
@@ -52,7 +56,7 @@ export class LanguagesComponent extends BaseListComponent<LanguageModel> impleme
    * @param {ListFormParams} parameters The current search parameters.
    * @returns {Observable<TEntity[]>}
    */
-  getAll(parameters?: ListFormParams){
+  getAll(parameters?: ListFormParams) {
     return this.languagesService.all(parameters);
   }
 
@@ -62,7 +66,7 @@ export class LanguagesComponent extends BaseListComponent<LanguageModel> impleme
    * @param {string} search The searching terms.
    * @returns {Observable<TotalModel>}
    */
-  getTotal(search?: string){
+  getTotal(search?: string) {
     return this.languagesService.allCount();
   }
 
@@ -72,7 +76,7 @@ export class LanguagesComponent extends BaseListComponent<LanguageModel> impleme
    * @param {LanguageModel} entity The current entity to delete.
    * @returns {Observable<any>}
    */
-  deleteEntity(entity: LanguageModel){
+  deleteEntity(entity: LanguageModel) {
     return this.languagesService.remove(entity.id);
   }
 }
