@@ -7,8 +7,9 @@ import { LanguagesService } from './languages.service';
 import { CreateComponent } from './create/create.component';
 import { EditComponent } from './edit/edit.component';
 import { CreateOrUpdateComponent } from './create-or-update/create-or-update.component';
-import { USERS_GUARD_PROVIDERS } from './guards';
 import { ResourcesComponent } from './resources/resources.component';
+import { USERS_GUARD_PROVIDERS } from './guards';
+import { LanguageResourcesDialogComponent } from './resources/resources-dialog.component';
 
 @NgModule({
   imports: [
@@ -20,11 +21,15 @@ import { ResourcesComponent } from './resources/resources.component';
     CreateComponent,
     EditComponent,
     CreateOrUpdateComponent,
+    LanguageResourcesDialogComponent,
     ResourcesComponent
   ],
   providers: [
     LanguagesService,
     USERS_GUARD_PROVIDERS
+  ],
+  entryComponents: [
+    LanguageResourcesDialogComponent,
   ]
 })
 export class LanguagesModule { }
