@@ -1,7 +1,6 @@
 import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { NgbModal, ModalDismissReasons, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
 import { TranslateService } from '@ngx-translate/core';
@@ -11,19 +10,22 @@ import 'rxjs/add/operator/debounceTime';
 import 'rxjs/add/operator/distinctUntilChanged';
 import 'rxjs/add/operator/do';
 import 'rxjs/add/operator/catch';
+import { MdDialog, MdDialogRef } from '@angular/material';
 
 import { LanguageModel } from '../language-model';
 import { ResourceModel } from '../resource-model';
 import { LanguagesService } from '../languages.service';
 import { environment } from '../../../../environments/environment';
-import { BaseListComponent, ListFormParams } from '../../../core/base-list-component';
+import { BaseListComponent } from '../../../core/base-list-component';
+import { ListFormParams } from '../../../core/list-form-params';
+import { LanguageResourcesDialogComponent } from './resources-dialog.component';
 
 @Component({
   selector: 'app-resources',
   templateUrl: './resources.component.html',
   styleUrls: ['./resources.component.scss']
 })
-export class ResourcesComponent extends BaseListComponent<ResourceModel> implements OnInit, OnDestroy {
+export class ResourcesComponent extends BaseListComponent<ResourceModel, ListFormParams> implements OnInit, OnDestroy {
 
   /** Gets or sets the language identifier @property {string} */
   languageId: string;
@@ -49,8 +51,6 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
   /** Gets or sets the child edit modal. @property {any} */
   @ViewChild('editModal') public editModal: any;
 
-  editModalRef: NgbModalRef;
-
   /**
    * Initializes a new instance of the ResourcesComponent.
    * @constructor
@@ -65,11 +65,12 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
     private route: ActivatedRoute,
     private router: Router,
     private service: LanguagesService,
+    public dialog: MdDialog,
+    public confirmDialog: MdDialog,
     formBuilder: FormBuilder,
-    modalService: NgbModal,
     private translateService: TranslateService
   ) {
-    super(formBuilder, modalService);
+    super(formBuilder, confirmDialog);
   }
 
   /**
@@ -77,8 +78,9 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
    * @method
    */
   ngOnInit() {
-    this.sorts.push({ column: 'key', direction: 'asc' });
     super.ngOnInit();
+    this.dataSource._sort.active = 'key';
+    this.dataSource._sort.direction = 'asc';
     this.initEditForm();
 
     this.routeSub = this.route.params.subscribe((params: Params) => {
@@ -89,7 +91,7 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
       }
 
       this.service.getById(this.languageId).subscribe(data => this.language = data);
-      this.initTable();
+
     });
   }
 
@@ -105,24 +107,12 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
   }
 
   /**
-   * Initializes the list table.
-   * @method
-   */
-  initTable() {
-    if (!this.languageId || this.languageId.replace(/ /ig, '') === ''){
-      return;
-    }
-
-    super.initTable();
-  }
-
-  /**
    * Gets all the element for a page.
    * @method
    * @param {ListFormParams} parameters The current search parameters.
    * @returns {Observable<TEntity[]>}
    */
-  getAll(parameters?: ListFormParams){
+  getAll(parameters?: ListFormParams) {
     return this.service.allResources(this.languageId, parameters);
   }
 
@@ -132,7 +122,7 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
    * @param {string} search The searching terms.
    * @returns {Observable<TotalModel>}
    */
-  getTotal(search?: string){
+  getTotal(search?: string) {
     return this.service.allResourcesCount(this.languageId, search);
   }
 
@@ -150,76 +140,11 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
   }
 
   /**
-   * Add a new resource to the language.
-   * @method
-   * @param {Event} event The current submit event.
-   */
-  addNew(event: Event) {
-    this.editResource(event, { id: '', key: '', value: '' });
-  }
-
-  /**
-   * Edit a resource to the language.
-   * @method
-   * @param {Event} event The current submit event.
-   * @param {ResourceModel} resource The resource model to edit.
-   */
-  editResource(event: Event, resource: ResourceModel) {
-    event.preventDefault();
-    this.action = !!resource.id ? 'edit' : 'create';
-    this.editForm.setValue(resource, { onlySelf: true });
-    this.showEditModal();
-  }
-
-  /**
-   * Save the editing resource.
-   * @method
-   * @param {Event} event The current submit event.
-   */
-  saveResource(event: Event) {
-    event.preventDefault();
-    this.isSubmitted = true;
-    if (!this.editForm.valid) {
-      return;
-    }
-
-    const isNew = !this.editForm.value.id;
-    const result = !isNew ? this.service.updateResource(this.languageId, this.editForm.value)
-      : this.service.createResource(this.languageId, this.editForm.value);
-
-    result.subscribe(data => {
-      this.initTable();
-      this.hideEditModal();
-
-      const lang = this.translateService.currentLang;
-      this.translateService.getTranslation(lang)
-        .subscribe(() => {
-          this.translateService.onLangChange.emit({ lang: lang, translations: result });
-        });
-    });
-  }
-
-  /**
-   * Show the edition modal.
-   * @method
-   */
-  public showEditModal(): void {
-    this.isSubmitted = false;
-    this.editModalRef = this.modalService.open(this.editModal)
-    this.editModalRef.result.then((result) => {
-      // save result
-    }, (reason) => {
-      // not saving
-    });
-  }
-
-  /**
    * Hide the edition modal.
    * @method
    */
   public hideEditModal(): void {
     this.isSubmitted = false;
-    this.editModalRef.close();
   }
 
   /**
@@ -228,8 +153,20 @@ export class ResourcesComponent extends BaseListComponent<ResourceModel> impleme
    * @param {ResourceModel} entity The current entity to delete.
    * @returns {Observable<any>}
    */
-  deleteEntity(entity: ResourceModel){
+  deleteEntity(entity: ResourceModel) {
     return this.service.removeResource(this.languageId, entity.id);
   }
 
+  public openDialog(resourceModel: ResourceModel) {
+    const dialogRef = this.dialog.open(LanguageResourcesDialogComponent);
+    dialogRef.componentInstance.languageId = this.languageId;
+
+    if (resourceModel !== undefined) {
+      dialogRef.componentInstance.setResourceModel(resourceModel);
+    }
+    dialogRef.afterClosed().subscribe(result => {
+      this.dataSource.refresh();
+      console.log(dialogRef.componentInstance.getResourceModel())
+    });
+  }
 }
