@@ -6,7 +6,6 @@ import { TranslateModule } from '@ngx-translate/core';
 import { Http } from '@angular/http';
 import { AuthHttp } from 'angular2-jwt';
 
-import { SharedModule } from '../shared/shared.module';
 import { LoginComponent } from './login/login.component';
 import { LogoutComponent } from './logout/logout.component';
 import { ForgotPasswordComponent } from './forgot-password/forgot-password.component';
@@ -15,6 +14,8 @@ import { AuthorizationService } from './authorization.service';
 import { authorizationHttpFactory } from './authorization-http';
 import { HasRightDirective } from './has-right.directive';
 import { USERS_GUARD_PROVIDERS } from './guards';
+import { MdButtonModule, MdCardModule, MdInputModule, MdIconModule } from '@angular/material';
+import { FlexLayoutModule } from '@angular/flex-layout';
 
 @NgModule({
   imports: [
@@ -23,6 +24,11 @@ import { USERS_GUARD_PROVIDERS } from './guards';
     ReactiveFormsModule,
     RouterModule,
     TranslateModule,
+    MdCardModule,
+    FlexLayoutModule,
+    MdInputModule,
+    MdButtonModule,
+    MdIconModule
   ],
   declarations: [
     HasRightDirective,
