@@ -1,6 +1,9 @@
 import { Component, OnInit, HostListener } from '@angular/core';
 
 import { AuthorizationService } from '../../authentication/authorization.service';
+import { UsersService } from '../../admin/users/users.service';
+
+import { UserModel } from '../../admin/users/user-model';
 
 @Component({
   selector: 'app-admin-layout',
@@ -15,6 +18,8 @@ export class AdminLayoutComponent implements OnInit {
 
   public disabled = false;
   public status: { isopen: boolean } = { isopen: false };
+
+ userModel = new UserModel();
 
   /**
    * Gets the current username.
@@ -31,7 +36,8 @@ export class AdminLayoutComponent implements OnInit {
    * @param {AuthorizationService} authorizationService The application authorization service.
    */
   constructor(
-    private authorizationService: AuthorizationService
+    private authorizationService: AuthorizationService,
+    private userService: UsersService
   ) { }
 
   public toggled(open: boolean): void {
@@ -41,11 +47,13 @@ export class AdminLayoutComponent implements OnInit {
   public toggleDropdown($event: MouseEvent): void {
     $event.preventDefault();
     $event.stopPropagation();
+    console.log('openclose');
     this.status.isopen = !this.status.isopen;
   }
 
   ngOnInit() {
     this.setWindowSize();
+    this.getcurrentUser();
   }
 
   /**
@@ -76,4 +84,14 @@ export class AdminLayoutComponent implements OnInit {
     return (this.windowHeight - margins) + 'px';
   }
 
+
+  /**
+   * Get the curent user
+   * @method
+   */
+  getcurrentUser() {
+    this.userService.getCurrentUser().subscribe((response) => {
+      this.userModel = response;
+    });
+  }
 }
